refactor: deduplicate connect_cluster_db into controlplane::connect()
Move the TiKV connection logic to nauka_hypervisor::controlplane::connect()
as a single public function. Remove identical copies from org, network,
and compute lib.rs files. All handlers now call the shared function.
